refactor: Payment 모델을 ERD 기준으로 통일

+++ src/main/java/com/back/sportteam/domain/payment/dto/response/PaymentPrepareResponse.java
@@ -4,7 +4,7 @@ import com.back.sportteam.domain.payment.entity.Payment;
 
 public record PaymentPrepareResponse(
         String merchantUid,
-        Long amount
+        Integer amount
 ) {
 
     public static PaymentPrepareResponse from(Payment payment) {
