+++ src/main/java/com/back/sportteam/domain/payment/service/PaymentAmountReader.java
@@ -2,7 +2,7 @@ package com.back.sportteam.domain.payment.service;
 
 public interface PaymentAmountReader {
 
-    Long getFacilityAmount(String matchId);
+    Integer getFacilityAmount(String matchId);
 
-    Long getParticipationAmount(String matchId);
+    Integer getParticipationAmount(String matchId);
 }
